style(siswa) menambhakan detail di home serta membuat route detail

+++ resources/views/siswa/index.blade.php
@@ -34,9 +34,9 @@
                     <img src="{{asset ('storage/' . $siswa->photo)}}" alt="Photo Siswa" width="50">
                 </td>
                 <td>
-                    <a href="/siswa/delete/{{$siswa->id}}">Delete</a> |
+                    <a onclick="return confirm('apakah mau di hapus?')" href="/siswa/delete/{{$siswa->id}}">Delete</a> |
                     <a href="">Edit</a> |
-                    <a href="">Detail</a> |
+                    <a href="/siswa/show/{{$siswa->id}}">Detail</a> |
                 </td>
             </tr>
             @endforeach
